2024-03-04 v0.34 Started work on timesone set screen
2024-03-04 v0.34 Started work on timesone set screen

+++ Firmware/DRM_Watch_v3/Preferences.ino
@@ -68,10 +68,10 @@ bool wifiSlotClear(int slot){ //1 ... wifiSlotCnt
 
 //----------------------//---------------------- TIMEZONE ------------//----------------------//----------------------//----------------------//----------------------
 
-int getTimeOffsetSec(){          //123456789012345
+long getTimeOffsetSec(){          //123456789012345
   return preferencesObject.getInt("timeOffsetSec", 60*60*2);
 }
-bool saveTimeOffsetSec(int value){
+bool saveTimeOffsetSec(long value){
   return preferencesObject.putInt("timeOffsetSec", value);
 }
